feat(server): initial commit; add database context; add users route;
Went ahead and set up an initial web API which already came with the initial route. Then - to make this a truly full-stack application - I added a connection to a Supabase Postgres database. Finally, I added a route to retrieve the data from the users table and return it to the frontend.